requirements.txt
boto==2.38.0
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.2
django-appconf==1.0.1
django-imagekit==3.2.6
django-storages==1.1.8
django-toolbelt==0.0.1
gunicorn==19.3.0
pilkit==1.1.12
Pillow==2.9.0
psycopg2==2.6.1
six==1.9.0
static3==0.6.1
wheel==0.24.0
